@@ -10,7 +10,9 @@ topics:
   - contracts
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.9.0 <4.0.0'
+
+resolution: workspace
 
 dev_dependencies:
-  lints: ^3.0.0
+  lints: ^4.0.0
