@@ -1,6 +1,6 @@
-import 'package:modularity_contracts/modularity_contracts.dart';
-import 'package:modularity_cli/modularity_cli.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:modularity_cli/modularity_cli.dart';
+import 'package:modularity_contracts/modularity_contracts.dart';
 
 // --- MOCK APP ARCHITECTURE ---
 
@@ -36,10 +36,9 @@ class CheckoutFeature extends Module {
 }
 
 class ProductDetailsFeature extends Module implements Configurable<int> {
-  late int productId;
-
   // Empty constructor for static graph
   ProductDetailsFeature();
+  late int productId;
 
   @override
   void configure(int args) {
@@ -53,8 +52,8 @@ class ProductDetailsFeature extends Module implements Configurable<int> {
 class ProductListFeature extends Module {
   @override
   List<Module> get submodules => [
-        ProductDetailsFeature(), // Composition: List owns Details
-      ];
+    ProductDetailsFeature(), // Composition: List owns Details
+  ];
 
   @override
   void binds(Binder i) {}
@@ -63,10 +62,10 @@ class ProductListFeature extends Module {
 class ShopFeature extends Module {
   @override
   List<Module> get submodules => [
-        ProductListFeature(),
-        CartFeature(),
-        CheckoutFeature(),
-      ];
+    ProductListFeature(),
+    CartFeature(),
+    CheckoutFeature(),
+  ];
 
   @override
   List<Module> get imports => [AnalyticsModule()];
@@ -86,16 +85,13 @@ class AccountFeature extends Module {
 class AppModule extends Module {
   @override
   List<Module> get submodules => [
-        ShopFeature(),
-        AccountFeature(),
-        AuthFeature(),
-      ];
+    ShopFeature(),
+    AccountFeature(),
+    AuthFeature(),
+  ];
 
   @override
-  List<Module> get imports => [
-        NetworkModule(),
-        AnalyticsModule(),
-      ];
+  List<Module> get imports => [NetworkModule(), AnalyticsModule()];
 
   @override
   void binds(Binder i) {}
