@@ -12,16 +12,17 @@ mixin _$CartStore on _CartStore, Store {
   Computed<int>? _$itemCountComputed;
 
   @override
-  int get itemCount => (_$itemCountComputed ??=
-          Computed<int>(() => super.itemCount, name: '_CartStore.itemCount'))
-      .value;
+  int get itemCount => (_$itemCountComputed ??= Computed<int>(
+    () => super.itemCount,
+    name: '_CartStore.itemCount',
+  )).value;
   Computed<double>? _$totalPriceComputed;
 
   @override
-  double get totalPrice =>
-      (_$totalPriceComputed ??= Computed<double>(() => super.totalPrice,
-              name: '_CartStore.totalPrice'))
-          .value;
+  double get totalPrice => (_$totalPriceComputed ??= Computed<double>(
+    () => super.totalPrice,
+    name: '_CartStore.totalPrice',
+  )).value;
 
   late final _$itemsAtom = Atom(name: '_CartStore.items', context: context);
 
@@ -38,13 +39,16 @@ mixin _$CartStore on _CartStore, Store {
     });
   }
 
-  late final _$_CartStoreActionController =
-      ActionController(name: '_CartStore', context: context);
+  late final _$_CartStoreActionController = ActionController(
+    name: '_CartStore',
+    context: context,
+  );
 
   @override
   void add(Product product) {
-    final _$actionInfo =
-        _$_CartStoreActionController.startAction(name: '_CartStore.add');
+    final _$actionInfo = _$_CartStoreActionController.startAction(
+      name: '_CartStore.add',
+    );
     try {
       return super.add(product);
     } finally {
@@ -54,8 +58,9 @@ mixin _$CartStore on _CartStore, Store {
 
   @override
   void remove(Product product) {
-    final _$actionInfo =
-        _$_CartStoreActionController.startAction(name: '_CartStore.remove');
+    final _$actionInfo = _$_CartStoreActionController.startAction(
+      name: '_CartStore.remove',
+    );
     try {
       return super.remove(product);
     } finally {
@@ -65,8 +70,9 @@ mixin _$CartStore on _CartStore, Store {
 
   @override
   void clear() {
-    final _$actionInfo =
-        _$_CartStoreActionController.startAction(name: '_CartStore.clear');
+    final _$actionInfo = _$_CartStoreActionController.startAction(
+      name: '_CartStore.clear',
+    );
     try {
       return super.clear();
     } finally {
