@@ -9,8 +9,10 @@ part of 'product_store.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic, no_leading_underscores_for_local_identifiers
 
 mixin _$ProductStore on _ProductStore, Store {
-  late final _$productsAtom =
-      Atom(name: '_ProductStore.products', context: context);
+  late final _$productsAtom = Atom(
+    name: '_ProductStore.products',
+    context: context,
+  );
 
   @override
   ObservableList<Product> get products {
@@ -25,8 +27,10 @@ mixin _$ProductStore on _ProductStore, Store {
     });
   }
 
-  late final _$isLoadingAtom =
-      Atom(name: '_ProductStore.isLoading', context: context);
+  late final _$isLoadingAtom = Atom(
+    name: '_ProductStore.isLoading',
+    context: context,
+  );
 
   @override
   bool get isLoading {
@@ -41,8 +45,10 @@ mixin _$ProductStore on _ProductStore, Store {
     });
   }
 
-  late final _$loadProductsAsyncAction =
-      AsyncAction('_ProductStore.loadProducts', context: context);
+  late final _$loadProductsAsyncAction = AsyncAction(
+    '_ProductStore.loadProducts',
+    context: context,
+  );
 
   @override
   Future<void> loadProducts() {
