@@ -1,9 +1,10 @@
 library modularity_contracts;
 
-export 'src/module.dart';
 export 'src/binder.dart';
 export 'src/binder_factory.dart';
-export 'src/logger.dart';
 export 'src/configurable.dart';
+export 'src/exceptions.dart';
+export 'src/logger.dart';
+export 'src/module.dart';
 export 'src/module_interceptor.dart';
 export 'src/retention_policy.dart';
