@@ -1,24 +1,20 @@
 import 'dart:async';
+
 import 'package:modularity_contracts/modularity_contracts.dart';
+
+import '../di/simple_binder.dart';
+import '../di/simple_binder_factory.dart';
 import '../graph/graph_resolver.dart';
 import '../graph/module_registry_key.dart';
-import '../di/simple_binder_factory.dart';
-import '../di/simple_binder.dart';
 import 'module_override_scope.dart';
 
-/// Контроллер, управляющий жизненным циклом одного модуля.
+/// Manage the full lifecycle of a single [Module]: resolution, initialisation,
+/// hot reload, and disposal.
 class ModuleController {
-  final Module module;
-  final Binder binder;
-  final BinderFactory _binderFactory; // Храним для создания импортов
-  final StreamController<ModuleStatus> _statusController;
-  final void Function(Binder)? overrides;
-  final ModuleOverrideScope? overrideScope;
-  final List<ModuleInterceptor> interceptors;
-
-  /// Ссылка на контроллеры импортируемых модулей.
-  final List<ModuleController> importedControllers = [];
-
+  /// Create a controller for [module] with optional DI configuration.
+  ///
+  /// When neither [binder] nor [binderFactory] is supplied, a default
+  /// [SimpleBinderFactory] is used.
   ModuleController(
     this.module, {
     Binder? binder,
@@ -26,26 +22,54 @@ class ModuleController {
     this.overrides,
     ModuleOverrideScope? overrideScopeTree,
     this.interceptors = const [],
-  })  : _statusController = StreamController<ModuleStatus>.broadcast(),
-        binder = binder ?? (binderFactory ?? SimpleBinderFactory()).create(),
-        _binderFactory = binderFactory ?? SimpleBinderFactory(),
-        overrideScope = overrideScopeTree?.withAdditionalOverride(overrides) ??
-            (overrides != null
-                ? ModuleOverrideScope(selfOverrides: overrides)
-                : overrideScopeTree) {
+  }) : _statusController = StreamController<ModuleStatus>.broadcast(),
+       binder = binder ?? (binderFactory ?? SimpleBinderFactory()).create(),
+       _binderFactory = binderFactory ?? SimpleBinderFactory(),
+       overrideScope =
+           overrideScopeTree?.withAdditionalOverride(overrides) ??
+           (overrides != null
+               ? ModuleOverrideScope(selfOverrides: overrides)
+               : overrideScopeTree) {
     _statusController.add(ModuleStatus.initial);
   }
 
+  /// The [Module] whose lifecycle this controller manages.
+  final Module module;
+
+  /// The [Binder] that holds all dependency registrations for [module].
+  final Binder binder;
+  final BinderFactory _binderFactory; // Храним для создания импортов
+  final StreamController<ModuleStatus> _statusController;
+
+  /// Optional callback applied to the [Binder] after binds/exports to override
+  /// registrations (e.g. for testing or feature flags).
+  final void Function(Binder)? overrides;
+
+  /// Hierarchical override scope propagated to imported modules.
+  final ModuleOverrideScope? overrideScope;
+
+  /// Ordered list of [ModuleInterceptor]s notified at each lifecycle event.
+  final List<ModuleInterceptor> interceptors;
+
+  /// Ссылка на контроллеры импортируемых модулей.
+  final List<ModuleController> importedControllers = [];
+
+  /// Broadcast stream of [ModuleStatus] transitions.
   Stream<ModuleStatus> get status => _statusController.stream;
   ModuleStatus _currentStatus = ModuleStatus.initial;
+
+  /// Return the most recent [ModuleStatus] of this controller.
   ModuleStatus get currentStatus => _currentStatus;
 
   RegistrationAwareBinder? get _registrationAwareBinder =>
       binder is RegistrationAwareBinder
-          ? binder as RegistrationAwareBinder
-          : null;
+      ? binder as RegistrationAwareBinder
+      : null;
 
   Object? _lastError;
+
+  /// Return the error captured during the last failed [initialize] call, or
+  /// `null` if no error occurred.
   Object? get lastError => _lastError;
 
   /// Конфигурация модуля.
@@ -56,9 +80,12 @@ class ModuleController {
       } catch (e) {
         // Handle generic type mismatch gracefully or rethrow
         // If we pass wrong type to configure(T args), Dart throws TypeError.
-        throw Exception("Module ${module.runtimeType} failed to configure: "
-            "Expected arguments of correct type for Configurable<T>.\n"
-            "Error: $e");
+        throw ModuleLifecycleException(
+          'Module ${module.runtimeType} failed to configure: '
+          'Expected arguments of correct type for Configurable<T>.\n'
+          'Error: $e',
+          moduleType: module.runtimeType,
+        );
       }
     }
   }
@@ -74,7 +101,9 @@ class ModuleController {
     }
 
     // Interceptor: onInit
-    for (var i in interceptors) i.onInit(module);
+    for (var i in interceptors) {
+      i.onInit(module);
+    }
 
     _updateStatus(ModuleStatus.loading);
 
@@ -102,16 +131,19 @@ class ModuleController {
         // Но на этом этапе Local пуст (binds еще не вызван).
         // Значит, мы проверяем Imports и Parent.
         if (!binder.contains(expectedType)) {
-          throw Exception(
-              "Module ${module.runtimeType} expects dependency of type '$expectedType', "
-              "but it was not found in Parent Scope or Imports.\n"
-              "Check if the parent module exports it or if it's correctly imported.");
+          throw ModuleConfigurationException(
+            "Module ${module.runtimeType} expects dependency of type '$expectedType', "
+            'but it was not found in Parent Scope or Imports.\n'
+            "Check if the parent module exports it or if it's correctly imported.",
+            moduleType: module.runtimeType,
+          );
         }
       }
 
       // 4. Binds (Private & Public)
-      final exportable =
-          binder is ExportableBinder ? binder as ExportableBinder : null;
+      final exportable = binder is ExportableBinder
+          ? binder as ExportableBinder
+          : null;
       exportable?.disableExportMode();
       module.binds(binder);
 
@@ -128,13 +160,17 @@ class ModuleController {
       _updateStatus(ModuleStatus.loaded);
 
       // Interceptor: onLoaded
-      for (var i in interceptors) i.onLoaded(module);
+      for (var i in interceptors) {
+        i.onLoaded(module);
+      }
     } catch (e) {
       _lastError = e;
       _updateStatus(ModuleStatus.error);
 
       // Interceptor: onError
-      for (var i in interceptors) i.onError(module, e);
+      for (var i in interceptors) {
+        i.onError(module, e);
+      }
 
       rethrow;
     }
@@ -151,8 +187,9 @@ class ModuleController {
     // В будущем SimpleBinder должен поддерживать "updateFactoryOnly".
 
     void rebind() {
-      final exportable =
-          binder is ExportableBinder ? binder as ExportableBinder : null;
+      final exportable = binder is ExportableBinder
+          ? binder as ExportableBinder
+          : null;
       exportable?.resetPublicScope();
       exportable?.disableExportMode();
       module.binds(binder);
@@ -165,12 +202,9 @@ class ModuleController {
 
     final aware = _registrationAwareBinder;
     if (aware != null) {
-      aware.runWithStrategy(
-        RegistrationStrategy.preserveExisting,
-        () {
-          rebind();
-        },
-      );
+      aware.runWithStrategy(RegistrationStrategy.preserveExisting, () {
+        rebind();
+      });
     } else {
       rebind();
     }
@@ -179,6 +213,7 @@ class ModuleController {
     module.hotReload(binder);
   }
 
+  /// Dispose of the module, its [Binder], and close the status stream.
   Future<void> dispose() async {
     _updateStatus(ModuleStatus.disposed);
     module.onDispose();
@@ -191,7 +226,9 @@ class ModuleController {
     await _statusController.close();
 
     // Interceptor: onDispose
-    for (var i in interceptors) i.onDispose(module);
+    for (var i in interceptors) {
+      i.onDispose(module);
+    }
   }
 
   void _updateStatus(ModuleStatus newStatus) {
