@@ -1,9 +1,12 @@
 import 'package:modularity_contracts/modularity_contracts.dart';
 
+/// Callback that applies dependency overrides to the given [Binder].
 typedef BinderOverride = void Function(Binder binder);
 
 BinderOverride? _composeOverrides(
-    BinderOverride? first, BinderOverride? second) {
+  BinderOverride? first,
+  BinderOverride? second,
+) {
   if (first == null) return second;
   if (second == null) return first;
   return (binder) {
@@ -12,15 +15,18 @@ BinderOverride? _composeOverrides(
   };
 }
 
-/// Дерево overrides для текущего модуля и его дочерних импортов.
+/// Hierarchical tree of [BinderOverride]s for the current module and its
+/// imported children.
 class ModuleOverrideScope {
-  final BinderOverride? selfOverrides;
-  final Map<Type, ModuleOverrideScope> children;
+  /// Create a scope with optional [selfOverrides] and child [children] map.
+  const ModuleOverrideScope({this.selfOverrides, this.children = const {}});
 
-  const ModuleOverrideScope({
-    this.selfOverrides,
-    this.children = const {},
-  });
+  /// Override callback applied to the owning module's [Binder], or `null` if
+  /// no overrides are needed at this level.
+  final BinderOverride? selfOverrides;
+
+  /// Per-module-type override scopes for imported (child) modules.
+  final Map<Type, ModuleOverrideScope> children;
 
   /// Возвращает скоуп для импортируемого модуля типа [type].
   ModuleOverrideScope? childFor(Type type) => children[type];
