@@ -25,8 +25,9 @@ class GraphResolver {
 
       // Check Circular Dependency (Immediate Fail-Fast)
       if (currentStack.contains(type)) {
-        throw Exception(
+        throw CircularDependencyException(
           'Circular dependency detected: ${currentStack.join(' -> ')} -> $type',
+          dependencyChain: [...currentStack, type],
         );
       }
 
@@ -63,14 +64,19 @@ class GraphResolver {
         // Если модуль уже грузится (его пнула другая ветка), просто ждем.
         // Проверяем на цикл именно в ЭТОЙ ветке
         if (currentStack.contains(type)) {
-          throw Exception(
-              'Circular dependency detected (during loading): ${currentStack.join(' -> ')} -> $type');
+          throw CircularDependencyException(
+            'Circular dependency detected (during loading): ${currentStack.join(' -> ')} -> $type',
+            dependencyChain: [...currentStack, type],
+          );
         }
         // "Smart Wait": Ждем пока другая ветка закончит работу
         await controller.status.firstWhere((s) => s == ModuleStatus.loaded);
       } else if (controller.currentStatus == ModuleStatus.error) {
-        throw Exception(
-            "Dependent module $type failed to load: ${controller.lastError}");
+        throw ModuleLifecycleException(
+          'Dependent module $type failed to load: ${controller.lastError}',
+          moduleType: type,
+          state: ModuleStatus.error,
+        );
       }
 
       return controller;
