@@ -1,14 +1,22 @@
 import 'package:flutter/widgets.dart';
 import 'package:modularity_core/modularity_core.dart';
 
+/// Inherited widget that exposes a [ModuleController] and its [Binder] to
+/// descendant widgets.
+///
+/// Typically inserted by [ModuleScope]; consumers obtain the [Binder] via
+/// [ModuleProvider.of] or retrieve the typed [Module] via
+/// [ModuleProvider.moduleOf].
 class ModuleProvider extends InheritedWidget {
-  final ModuleController controller;
-
+  /// Create a [ModuleProvider] that exposes [controller] to [child].
   const ModuleProvider({
-    Key? key,
+    super.key,
     required this.controller,
-    required Widget child,
-  }) : super(key: key, child: child);
+    required super.child,
+  });
+
+  /// Controller that owns the DI [Binder] and the [Module] lifecycle.
+  final ModuleController controller;
 
   @override
   bool updateShouldNotify(ModuleProvider oldWidget) {
@@ -25,25 +33,32 @@ class ModuleProvider extends InheritedWidget {
         : context.getInheritedWidgetOfExactType<ModuleProvider>();
 
     if (provider == null) {
-      throw Exception('ModuleProvider not found in context');
+      throw ModuleConfigurationException(
+        'ModuleProvider not found in context.',
+      );
     }
     return provider.controller.binder;
   }
 
   /// Получает сам модуль типа [M] из контекста.
-  static M moduleOf<M extends Module>(BuildContext context,
-      {bool listen = true}) {
+  static M moduleOf<M extends Module>(
+    BuildContext context, {
+    bool listen = true,
+  }) {
     final provider = listen
         ? context.dependOnInheritedWidgetOfExactType<ModuleProvider>()
         : context.getInheritedWidgetOfExactType<ModuleProvider>();
 
     if (provider == null) {
-      throw Exception('ModuleProvider not found in context');
+      throw ModuleConfigurationException(
+        'ModuleProvider not found in context.',
+      );
     }
 
     if (provider.controller.module is! M) {
-      throw Exception(
-          'Nearest module is ${provider.controller.module.runtimeType}, but expected $M');
+      throw ModuleConfigurationException(
+        'Nearest module is ${provider.controller.module.runtimeType}, but expected $M.',
+      );
     }
 
     return provider.controller.module as M;
